* autogen.sh: Do not try to delete nonexistant files. 	* util/import_gcrypth.sed: Add some missing header removals.

+++ autogen.sh
@@ -14,13 +14,19 @@ python util/import_unicode.py unicode/UnicodeData.txt unicode/BidiMirroring.txt 
 echo "Importing libgcrypt..."
 python util/import_gcry.py grub-core/lib/libgcrypt/ grub-core
 sed -n -f util/import_gcrypth.sed < grub-core/lib/libgcrypt/src/gcrypt.h.in > include/grub/gcrypt/gcrypt.h
-rm include/grub/gcrypt/g10lib.h
-rm -rf grub-core/lib/libgcrypt-grub/mpi/generic
+if [ -f include/grub/gcrypt/g10lib.h ]; then
+    rm include/grub/gcrypt/g10lib.h
+fi
+if [ -d grub-core/lib/libgcrypt-grub/mpi/generic ]; then 
+    rm -rf grub-core/lib/libgcrypt-grub/mpi/generic
+fi
 ln -s ../../../grub-core/lib/libgcrypt-grub/src/g10lib.h include/grub/gcrypt/g10lib.h
 cp -R grub-core/lib/libgcrypt/mpi/generic grub-core/lib/libgcrypt-grub/mpi/generic
 
 for x in mpi-asm-defs.h mpih-add1.c mpih-sub1.c mpih-mul1.c mpih-mul2.c mpih-mul3.c mpih-lshift.c mpih-rshift.c; do
-    rm grub-core/lib/libgcrypt-grub/mpi/"$x"
+    if [ -f grub-core/lib/libgcrypt-grub/mpi/"$x" ]; then
+	rm grub-core/lib/libgcrypt-grub/mpi/"$x"
+    fi
     ln -s generic/"$x" grub-core/lib/libgcrypt-grub/mpi/"$x"
 done
